+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* remove deprecated web endpoints for base-items, components, workflows, wirings and documentation
+* remove exporting capabilities for exporting from the pre 0.7 Java backend
+
 ## 0.12.1
 * dependency upgrades and upgrade to Python 3.13
 * **BREAKING CHANGE**: removed u8darts from runtime image due to non-maintained transitive dependencies